Reverse refactor code for the next PR
